+++ aiida/orm/data/base.py
@@ -8,18 +8,41 @@
 # For further information please visit http://www.aiida.net               #
 ###########################################################################
 from abc import ABCMeta
-import numbers
-import collections
+from collections import MutableSequence
 
 try:
     from functools import singledispatch
 except ImportError:
     from singledispatch import singledispatch
-from past.builtins import basestring
-import numpy as np
 
 from aiida.orm import Data
-from aiida.orm.data.parameter import ParameterData 
+
+
+@singledispatch
+def to_aiida_type(value):
+    """
+    Turns basic Python types (str, int, float, bool) into the corresponding AiiDA types.
+    """
+    raise TypeError("Cannot convert value of type {} to AiiDA type.".format(type(value)))
+
+
+def _left_operator(func):
+    def inner(self, other):
+        l = self.value
+        if isinstance(other, NumericType):
+            r = other.value
+        else:
+            r = other
+        return to_aiida_type(func(l, r))
+    return inner
+
+
+def _right_operator(func):
+    def inner(self, other):
+        assert not isinstance(other, NumericType)
+        return to_aiida_type(func(self.value, other))
+    return inner
+
 
 class BaseType(Data):
     """
@@ -97,21 +120,6 @@ class BaseType(Data):
 
         return kwargs
 
-def _left_operator(func):
-    def inner(self, other):
-        l = self.value
-        if isinstance(other, NumericType):
-            r = other.value
-        else:
-            r = other
-        return to_aiida_type(func(l, r))
-    return inner
-
-def _right_operator(func):
-    def inner(self, other):
-        assert not isinstance(other, NumericType)
-        return to_aiida_type(func(self.value, other))
-    return inner
 
 class NumericType(BaseType):
     """
@@ -169,46 +177,7 @@ class NumericType(BaseType):
         return int(self.value)
 
 
-class Float(NumericType):
-    """
-    Class to store float numbers as AiiDA nodes
-    """
-    _type = float
-
-
-class Int(NumericType):
-    """
-    Class to store integer numbers as AiiDA nodes
-    """
-    _type = int
-
-
-class Str(BaseType):
-    """
-    Class to store strings as AiiDA nodes
-    """
-    _type = str
-
-
-class Bool(BaseType):
-    """
-    Class to store booleans as AiiDA nodes
-    """
-    _type = bool
-
-    def __int__(self):
-        return int(bool(self))
-
-    # Python 2
-    def __nonzero__(self):
-        return self.__bool__()
-
-    # Python 3
-    def __bool__(self):
-        return self.value
-
-
-class List(Data, collections.MutableSequence):
+class List(Data, MutableSequence):
     """
     Class to store python lists as AiiDA nodes
     """
@@ -307,57 +276,3 @@ class List(Data, collections.MutableSequence):
         :rtype: bool
         """
         return not self.is_stored
-
-
-def get_true_node():
-    """
-    Return a Bool Data node, with value True
-
-    Cannot be done as a singleton in the module, because it would be generated
-    at import time, with the risk that (e.g. in the tests, or at the very first use
-    of AiiDA) a user is not yet defined in the DB (but a user is mandatory in the
-    DB before you can create new Nodes in AiiDA).
-    """
-    TRUE = Bool(typevalue=(bool, True))
-    return TRUE
-
-
-def get_false_node():
-    """
-    Return a Bool Data node, with value False
-
-    Cannot be done as a singleton in the module, because it would be generated
-    at import time, with the risk that (e.g. in the tests, or at the very first use
-    of AiiDA) a user is not yet defined in the DB (but a user is mandatory in the
-    DB before you can create new Nodes in AiiDA).
-    """
-    FALSE = Bool(typevalue=(bool, False))
-    return FALSE
-
-@singledispatch
-def to_aiida_type(value):
-    """
-    Turns basic Python types (str, int, float, bool) into the corresponding AiiDA types.
-    """
-    raise TypeError("Cannot convert value of type {} to AiiDA type.".format(type(value)))
-
-@to_aiida_type.register(basestring)
-def _(value):
-    return Str(value)
-
-@to_aiida_type.register(numbers.Integral)
-def _(value):
-    return Int(value)
-
-@to_aiida_type.register(numbers.Real)
-def _(value):
-    return Float(value)
-
-@to_aiida_type.register(bool)
-@to_aiida_type.register(np.bool_)
-def _(value):
-    return Bool(value)
-
-@to_aiida_type.register(dict)
-def _(value):
-    return ParameterData(dict=value)
